feat(navbar): responsive navbar both large and mobile devices
