+++ src/data/navigation.js
@@ -1,0 +1,10 @@
+// Navigation links data
+export const navLinks = [
+  { path: "/", label: "Home" },
+  { path: "/seafood", label: "Seafood" },
+  { path: "/souvenirs", label: "Souvenirs" },
+  { path: "/about", label: "About" },
+  // { path: "/cart", label: "Cart" },
+  // { path: "/profile", label: "Profile" },
+  // { path: "/colors", label: "Colors" },
+];
