+++ src/index.css
@@ -1,5 +1,6 @@
 @import url("https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700&family=Playfair+Display:wght@400;500;600;700&display=swap");
 @import "tailwindcss";
+@plugin "daisyui";
 @plugin "daisyui/theme" {
   name: "coast2cart";
   default: true;
